refactor: 폴더 구조 변경
point 폴더 내부에서 관련 파일이 추가로 생성될 여지가 있는 파일은 관심사 별로 별도 폴더를 두어 관리

- point model 위치
  - /point -> /point/model
- point dto 위치
  - /point -> /point/dto
- point spec 위치
  - /point -> /point/spec
