+++ src/database/userpoint/userpoint.table.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { randomInt } from 'crypto';
-import { UserPoint } from 'src/point/point.model';
+import { UserPoint } from 'src/point/model/point.model';
 
 /**
  * 해당 Table 클래스는 변경하지 않고 공개된 API 만을 사용해 데이터를 제어합니다.
