+++ src/point/model/point.model.ts
